shuffle some stuff around
