+++ vowpalwabbit/example_predict.cc
@@ -58,16 +58,3 @@ std::string features_to_string(const example_predict& ec)
   }
   return strstream.str();
 }
-
-std::string debug_depth_indent_string(const int32_t depth)
-{
-  constexpr const char* indent_str = "- ";
-  constexpr const char* space_str = "  ";
-
-  if (depth == 0) return indent_str;
-
-  std::stringstream str_stream;
-  for (int32_t i = 0; i < depth - 1; i++) { str_stream << space_str; }
-  str_stream << indent_str;
-  return str_stream.str();
-}
